chore(logger): make directory and move Context
