+++ packages/@ndxbn/logger/src/Context/ContextGeneratorInterface.ts
@@ -1,0 +1,7 @@
+export default interface ContextGeneratorInterface {
+  /**
+   * context name
+   */
+  readonly name: string;
+  generate(): import("./types").Context;
+}
