+++ packages/@ndxbn/logger/src/Context/types.ts
@@ -1,0 +1,3 @@
+export type Context = {
+  [k: string]: { toString(): string };
+};
